Removed a bunch of exports
I’d like to keep the external API as simple as possible for now.  The
internal functions are documented, and users are free to find and use
them, but I don’t want to make the promise that these functions will be
there in future releases.  I want the flexibility to re-work the
internal API.

We can release the internal API piece by piece as it matures and
solidifies.