Makes all CaretList's predictions numeric
Caret not always returns numeric predictions, instead sometimes it returns them as character. The underlying packages that implement the models themselves are in constant change and cannot be 100% trusted to always give the numeric values CaretEnsemble and CaretStack expect. In order to solve this issue, the line  if(!is.null(model)) model$pred$pred <- as.numeric(model$pred$pred) can be added on the caretList function.